refactor(common): move `FoundryTransactionBuilder` to `foundry-common`
