+++ crates/cheatcodes/Cargo.toml
@@ -20,7 +20,6 @@ foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
 foundry-evm-core.workspace = true
-foundry-primitives.workspace = true
 foundry-evm-fuzz.workspace = true
 foundry-evm-traces.workspace = true
 foundry-wallets.workspace = true
