+++ crates/common/Cargo.toml
@@ -17,7 +17,6 @@ foundry-block-explorers = { workspace = true, features = ["foundry-compilers"] }
 foundry-common-fmt.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
-foundry-primitives.workspace = true
 
 alloy-chains.workspace = true
 alloy-dyn-abi = { workspace = true, features = ["arbitrary", "eip712"] }
@@ -79,6 +78,8 @@ anstyle.workspace = true
 ciborium.workspace = true
 
 flate2.workspace = true
+tempo-alloy.workspace = true
+tempo-primitives.workspace = true
 
 [build-dependencies]
 chrono.workspace = true
