+++ crates/common/src/transactions/receipt.rs
@@ -1,17 +1,11 @@
-//! Wrappers for transactions.
-
-use alloy_consensus::Transaction;
-use alloy_eips::eip7702::SignedAuthorization;
-use alloy_network::{AnyTransactionReceipt, Network, TransactionResponse};
-use alloy_primitives::{Address, Bytes, U256};
+use alloy_network::{AnyNetwork, AnyTransactionReceipt, Network, TransactionResponse};
 use alloy_provider::{
     Provider,
-    network::{AnyNetwork, ReceiptResponse, TransactionBuilder},
+    network::{ReceiptResponse, TransactionBuilder},
 };
 use alloy_rpc_types::BlockId;
 use eyre::Result;
 use foundry_common_fmt::{UIfmt, UIfmtReceiptExt, get_pretty_receipt_attr};
-use foundry_primitives::FoundryTransactionBuilder;
 use serde::{Deserialize, Serialize};
 
 /// Helper type to carry a transaction along with an optional revert reason
@@ -92,49 +86,6 @@ revertReason         {}",
     }
 }
 
-impl<N: Network> UIfmt for TransactionMaybeSigned<N>
-where
-    N::TxEnvelope: UIfmt,
-    N::TransactionRequest: FoundryTransactionBuilder<N>,
-{
-    fn pretty(&self) -> String {
-        match self {
-            Self::Signed { tx, .. } => tx.pretty(),
-            Self::Unsigned(tx) => format!(
-                "
-accessList           {}
-chainId              {}
-gasLimit             {}
-gasPrice             {}
-input                {}
-maxFeePerBlobGas     {}
-maxFeePerGas         {}
-maxPriorityFeePerGas {}
-nonce                {}
-to                   {}
-type                 {}
-value                {}",
-                tx.access_list()
-                    .as_ref()
-                    .map(|a| a.iter().collect::<Vec<_>>())
-                    .unwrap_or_default()
-                    .pretty(),
-                tx.chain_id().pretty(),
-                tx.gas_limit().unwrap_or_default(),
-                tx.gas_price().pretty(),
-                tx.input().pretty(),
-                tx.max_fee_per_blob_gas().pretty(),
-                tx.max_fee_per_gas().pretty(),
-                tx.max_priority_fee_per_gas().pretty(),
-                tx.nonce().pretty(),
-                tx.to().pretty(),
-                tx.output_tx_type(),
-                tx.value().pretty(),
-            ),
-        }
-    }
-}
-
 fn extract_revert_reason<S: AsRef<str>>(error_string: S) -> Option<String> {
     let message_substr = "execution reverted: ";
     error_string
@@ -157,91 +108,6 @@ where
         return Some(receipt.revert_reason.pretty());
     }
     get_pretty_receipt_attr::<N>(&receipt.receipt, attr)
-}
-
-/// Used for broadcasting transactions
-/// A transaction can either be a `TransactionRequest` waiting to be signed
-/// or a `TxEnvelope`, already signed
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum TransactionMaybeSigned<N: Network> {
-    Signed {
-        #[serde(flatten)]
-        tx: N::TxEnvelope,
-        from: Address,
-    },
-    Unsigned(N::TransactionRequest),
-}
-
-impl<N: Network> TransactionMaybeSigned<N> {
-    /// Creates a new (unsigned) transaction for broadcast
-    pub fn new(tx: N::TransactionRequest) -> Self {
-        Self::Unsigned(tx)
-    }
-
-    pub fn is_unsigned(&self) -> bool {
-        matches!(self, Self::Unsigned(_))
-    }
-
-    pub fn as_unsigned_mut(&mut self) -> Option<&mut N::TransactionRequest> {
-        match self {
-            Self::Unsigned(tx) => Some(tx),
-            _ => None,
-        }
-    }
-
-    pub fn from(&self) -> Option<Address> {
-        match self {
-            Self::Signed { from, .. } => Some(*from),
-            Self::Unsigned(tx) => tx.from(),
-        }
-    }
-
-    pub fn input(&self) -> Option<&Bytes> {
-        match self {
-            Self::Signed { tx, .. } => Some(tx.input()),
-            Self::Unsigned(tx) => tx.input(),
-        }
-    }
-
-    pub fn to(&self) -> Option<Address> {
-        match self {
-            Self::Signed { tx, .. } => tx.to(),
-            Self::Unsigned(tx) => tx.to(),
-        }
-    }
-
-    pub fn value(&self) -> Option<U256> {
-        match self {
-            Self::Signed { tx, .. } => Some(tx.value()),
-            Self::Unsigned(tx) => tx.value(),
-        }
-    }
-
-    pub fn gas(&self) -> Option<u128> {
-        match self {
-            Self::Signed { tx, .. } => Some(tx.gas_limit() as u128),
-            Self::Unsigned(tx) => tx.gas_limit().map(|g| g as u128),
-        }
-    }
-
-    pub fn nonce(&self) -> Option<u64> {
-        match self {
-            Self::Signed { tx, .. } => Some(tx.nonce()),
-            Self::Unsigned(tx) => tx.nonce(),
-        }
-    }
-
-    pub fn authorization_list(&self) -> Option<Vec<SignedAuthorization>>
-    where
-        N::TransactionRequest: FoundryTransactionBuilder<N>,
-    {
-        match self {
-            Self::Signed { tx, .. } => tx.authorization_list().map(|auths| auths.to_vec()),
-            Self::Unsigned(tx) => tx.authorization_list().cloned(),
-        }
-        .filter(|auths| !auths.is_empty())
-    }
 }
 
 #[cfg(test)]
