+++ crates/forge/Cargo.toml
@@ -64,7 +64,6 @@ alloy-provider = { workspace = true, features = ["reqwest", "ws", "ipc"] }
 alloy-signer.workspace = true
 alloy-transport.workspace = true
 
-foundry-primitives.workspace = true
 tempo-alloy.workspace = true
 
 revm.workspace = true
