+++ crates/primitives/src/network/mod.rs
@@ -1,13 +1,11 @@
 use alloy_network::Network;
 
 mod receipt;
-mod transaction;
 
 use alloy_provider::fillers::{
     BlobGasFiller, ChainIdFiller, GasFiller, JoinFill, NonceFiller, RecommendedFillers,
 };
 pub use receipt::*;
-pub use transaction::FoundryTransactionBuilder;
 
 /// Foundry network type.
 ///
