+++ crates/script/Cargo.toml
@@ -24,7 +24,6 @@ foundry-debugger.workspace = true
 foundry-cheatcodes.workspace = true
 foundry-wallets.workspace = true
 foundry-linking.workspace = true
-foundry-primitives.workspace = true
 forge-script-sequence.workspace = true
 
 serde.workspace = true
